backend/requirements.txt
annotated-types==0.6.0
anyio==4.0.0
bcrypt==4.0.1
certifi==2023.11.17
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.1.3
fastapi==0.104.1
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
nodeenv==1.8.0
passlib==1.7.4
prisma==0.11.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.5.1
pydantic-core==2.14.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
tomlkit==0.12.3
typing-extensions==4.8.0
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
